Move sample_requests and generate_synthetic_dataset to commons
